requirement.txt
absl-py==1.0.0
alabaster==0.7.12
aniso8601==9.0.1
anyio==3.6.1
appdirs==1.4.4
astroid==2.5.2
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
black==20.8b1
build==0.8.0
cachetools==5.0.0
cachey==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
dask==2022.8.0
debugpy==1.6.0
decorator==5.1.1
docstring-parser==0.14.1
docutils==0.19
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==2.0
fonttools==4.33.3
freetype-py==2.3.0
fsspec==2022.7.1
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.6.0
HeapDict==1.0.1
hsluv==5.0.3
httptools==0.4.0
idna==3.3
imageio==2.21.1
imagesize==1.4.1
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.10.0
jupyter-client==7.3.1
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.6.0
libclang==14.0.1
locket==1.0.0
magicgui==0.5.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
napari==0.4.16
napari-console==0.0.4
napari-plugin-engine==0.2.0
napari-svg==0.1.6
nest-asyncio==1.5.5
networkx==2.8.5
npe2==0.6.1
numpy==1.22.3
numpydoc==1.4.0
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
partd==1.3.0
pathspec==0.8.1
pep517==0.13.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.1.0
Pint==0.19.2
plotly==5.10.0
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
psygnal==0.3.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
Pygments==2.12.0
pylint==2.7.4
PyOpenGL==3.1.6
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytomlpp==1.0.11
pytz==2022.1
PyWavelets==1.3.0
pywin32==304
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.1
QtPy==2.2.0
regex==2021.4.4
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.8
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.2.0
starlette==0.19.1
superqt==0.3.3
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
typed-ast==1.4.2
typer==0.6.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.1
vispy==0.10.0
watchfiles==0.15.0
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.0
